+++ addOns/authhelper/src/main/java/org/zaproxy/addon/authhelper/AuthTestDialog.java
@@ -33,6 +33,7 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
+import java.util.regex.Pattern;
 import javax.swing.BorderFactory;
 import javax.swing.Icon;
 import javax.swing.ImageIcon;
@@ -113,6 +114,7 @@ public class AuthTestDialog extends StandardFieldsDialog {
     private static final String RECORD_DIAGNOSTICS_LABEL =
             "authhelper.auth.test.dialog.label.recdiag";
     private static final String DIAGNOSTICS_LABEL = "authhelper.auth.test.dialog.label.diag";
+    private static final String DOMAINS_LABEL = "authhelper.auth.test.dialog.label.domains";
     private static final String COPY_LABEL = "authhelper.auth.test.dialog.label.copy";
 
     private static final String FOUND_STR =
@@ -163,6 +165,7 @@ public class AuthTestDialog extends StandardFieldsDialog {
                 DisplayUtils.getScaledDimension(600, 550),
                 new String[] {
                     "authhelper.auth.test.dialog.tab.test",
+                    "authhelper.auth.test.dialog.tab.domains",
                     "authhelper.auth.test.dialog.tab.steps",
                     "authhelper.auth.test.dialog.tab.diag"
                 });
@@ -251,6 +254,10 @@ public class AuthTestDialog extends StandardFieldsDialog {
         this.addPadding(0);
 
         int tab = 1;
+
+        addMultilineField(tab, DOMAINS_LABEL, params.getDomains());
+
+        tab++;
         stepsPanel = new StepsPanel(this, true);
         stepsPanel.setSteps(params.getSteps());
         setCustomTabPanel(tab, stepsPanel.getPanel());
@@ -467,6 +474,14 @@ public class AuthTestDialog extends StandardFieldsDialog {
             String loginUrl = this.getStringValue(LOGIN_URL_LABEL);
             context.addIncludeInContextRegex(
                     SessionStructure.getHostName(new URI(loginUrl, false)) + ".*");
+
+            for (String dom : getDomains()) {
+                if (dom.endsWith(".*")) {
+                    // Just in case the user has added this anyway
+                    dom = dom.substring(0, dom.length() - 2);
+                }
+                context.addIncludeInContextRegex(Pattern.quote(dom) + ".*");
+            }
 
             JComboBox<?> browserCombo = (JComboBox<?>) this.getField(BROWSER_LABEL);
             String browserId = ((BrowserUI) browserCombo.getSelectedItem()).getBrowser().getId();
@@ -692,6 +707,12 @@ public class AuthTestDialog extends StandardFieldsDialog {
         }
     }
 
+    private List<String> getDomains() {
+        return List.of(this.getStringValue(DOMAINS_LABEL).split("\r?\n")).stream()
+                .filter(StringUtils::isNotBlank)
+                .toList();
+    }
+
     private static void setTotp(
             List<AuthenticationStep> steps, UsernamePasswordAuthenticationCredentials credentials) {
         if (!TotpSupport.isTotpInCore()) {
@@ -755,6 +776,7 @@ public class AuthTestDialog extends StandardFieldsDialog {
                                         Constant.messages.getString(
                                                 "authhelper.auth.test.dialog.default-context"));
                         setFieldValue(LOGIN_URL_LABEL, "");
+                        setFieldValue(DOMAINS_LABEL, "");
                         setFieldValue(USERNAME_LABEL, "");
                         setFieldValue(PASSWORD_LABEL, "");
                         setFieldValue(LOGIN_WAIT_LABEL, AuthhelperParam.DEFAULT_WAIT);
@@ -786,6 +808,7 @@ public class AuthTestDialog extends StandardFieldsDialog {
     private void saveDetails() {
         AuthhelperParam params = this.ext.getParam();
         params.setLoginUrl(this.getStringValue(LOGIN_URL_LABEL));
+        params.setDomains(this.getStringValue(DOMAINS_LABEL));
         params.setUsername(this.getStringValue(USERNAME_LABEL));
         JComboBox<?> browserCombo = (JComboBox<?>) this.getField(BROWSER_LABEL);
         params.setBrowser(((BrowserUI) browserCombo.getSelectedItem()).getBrowser().getId());
@@ -813,6 +836,17 @@ public class AuthTestDialog extends StandardFieldsDialog {
             }
             if (this.getStringValue(PASSWORD_LABEL).isBlank()) {
                 return Constant.messages.getString("authhelper.auth.test.dialog.error.nopassword");
+            }
+        }
+        for (String dom : this.getDomains()) {
+            String domLc = dom.toLowerCase(Locale.ROOT);
+            if (!domLc.startsWith("http://") && !domLc.startsWith("https://")) {
+                return Constant.messages.getString("authhelper.auth.test.dialog.error.baddom", dom);
+            }
+            try {
+                new URI(dom, false);
+            } catch (Exception e) {
+                return Constant.messages.getString("authhelper.auth.test.dialog.error.baddom", dom);
             }
         }
         return null;